Update PerspectiveTransform.php
On some error correction levels the $denominator variable on line 114 will end up being the integer 0. This is then used as the denominator in the next 2 lines in division which will cause PHP to fatal error and crash the program. By throwing an exception instead we can catch the error and handle it properly.